My_app/requirements.txt
wxPython>=4.2.0
numpy>=1.24.0
Pillow>=10.0.0
pandas>=2.0.0
scikit-image>=0.21.0
scipy>=1.10.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
scipy>=1.10.0
scikit-image>=0.21.0
